backend/requirements.txt
django
gunicorn==22.0.0
psycopg2
djangorestframework
djangorestframework-simplejwt
django-cors-headers